Fonctionnement de la <- de l'AppBar
