+++ app/src/main/java/com/insa/mygamelist/MainActivity.kt
@@ -7,21 +7,27 @@ import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
+import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.material3.Text
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.toRoute
 import com.insa.mygamelist.data.IGDB
-import com.insa.mygamelist.ui.theme.GameDetail
-import com.insa.mygamelist.ui.theme.GameScreen
-import com.insa.mygamelist.ui.theme.Home
-import com.insa.mygamelist.ui.theme.ListOfGames
+import com.insa.mygamelist.ui.myAppBar
+import com.insa.mygamelist.ui.navigation.GameDetail
+import com.insa.mygamelist.ui.views.GameScreen
+import com.insa.mygamelist.ui.navigation.Home
+import com.insa.mygamelist.ui.views.ListOfGames
 import com.insa.mygamelist.ui.theme.MyGamesListTheme
 
 class MainActivity : ComponentActivity() {
 
+    @OptIn(ExperimentalMaterial3Api::class)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
@@ -32,8 +38,9 @@ class MainActivity : ComponentActivity() {
             val navController = rememberNavController()
 
             MyGamesListTheme {
-                Scaffold(
-                    modifier = Modifier.fillMaxSize()) { innerPadding ->
+                Scaffold(topBar = {
+                    myAppBar(navController)
+                }, modifier = Modifier.fillMaxSize()) { innerPadding ->
                     NavHost(navController = navController, startDestination = Home) {
                         composable<Home> {
                             ListOfGames(
