+++ notes.md
@@ -1,0 +1,10 @@
+### Documentations utilisées :
+
+
+#### Navigation Component
+https://developer.android.com/guide/navigation/design?hl=fr#compose
+
+https://developer.android.com/guide/navigation/design/type-safety?hl=fr
+
+
+#### AppBar
